Move tooltip component; appease TS
