+++ packages/frontend-shared/src/gql-components/topnav/UnsupportedBrowserTooltip.vue
